Updated Code.
The php_plays_txt.html was renamed to index.html. Renamed ph_styles.css. In the ph_syles.css the charset was set to utf-8. Fontfaces for Champagne, Grunge, and Dobkin. The background color of the html to be set to greenish gray. The body background-color was set to whiteish. The font family was set to Palatino Lino Type for the body. The header was given a background color of black. The margin and the padding of the paragraph was adjusted. The font size and text-align was also adjusted for paragraphs directly nested in the body. The navigation links were given a font, and padding top/bottom. The styles for hover was changed to remove the underline and was given color. The individual playbill were given a body color. The h1 of those bills were given a specific font family to use. The description list were formatted.